requirements.txt
beautifulsoup4==4.10.0
numpy==1.21.3
regex==2022.10.31
requests==2.27.1
retrying==1.3.3
tqdm==4.65.0
mysql-connector==2.2.9